pyproject.toml
[project]
name = "trm-vs-mdlm"
version = "0.1.0"
description = "TRM vs MDLM training comparison for masked language modeling"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "datasets>=2.14.0",
    "accelerate>=0.24.0",
    "wandb>=0.15.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
]
